requirements.txt
Flask==3.00
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tensorflow==2.13.0
gunicorn
streamlit==1.28.0
pillow==10.0.0
